port: fix LP64 long→8-byte type bugs (chara structs, model matrices, VabHdr, ProgAtr)
PSX idioms assumed long = 4 bytes. On macOS LP64 (long = 8) this corrupted
NPC positions and model matrices; Windows LLP64 (long = 4) was unaffected.

- NPC movement (alessa, bloody_incubator, bloody_lisa, cheryl, cybil, dahlia,
  ghost_child_alessa, incubator, kaufmann, lisa): the pattern
    offset.vx = (u32)Q12_MULT_PRECISE(...) << shift
  stored a u32 into VECTOR3::vx (long). On LP64 the u32 is zero-extended,
  turning negative sin/cos values into large positive offsets; after a few
  frames the NPC drifts off the grid, Ipd_CollisionDataGet returns NULL, and
  the NPC falls through the floor. Wrap in (s32)(...) so the 32-bit pattern
  is sign-extended into long.
- ipd_reformat.c ParseIpdModelBufferC: memcpy(&dst->mat, &src[4], sizeof(MATRIX))
  copied 48 bytes (long t[3] widens to 24 bytes + alignment pad on LP64) from
  a 32-byte PSX binary MATRIX, overrunning 16 bytes and corrupting mat.t[].
  Parse m[3][3] and t[3] field-by-field at their known PSX binary offsets.
- libsnd.h: VabHdr/ProgAtr used long members; switch to s32/u32 so the structs
  match the PSX binary layout on LP64.

+++ pc_port/include/psyq/libsnd.h
@@ -59,11 +59,14 @@
 #endif
 
 /* VAB structures */
+/* long is 8 bytes on LP64 (macOS/Linux) but 4 bytes on PSX and Windows LLP64.
+ * These structs map directly to binary disc data (PSX 4-byte fields), so use
+ * fixed-width types to match the on-disc layout on all platforms. */
 typedef struct VabHdr {
-    long           form;
-    long           ver;
-    long           id;
-    unsigned long  fsize;
+    s32            form;
+    s32            ver;
+    s32            id;
+    u32            fsize;
     unsigned short reserved0;
     unsigned short ps;
     unsigned short ts;
@@ -72,7 +75,7 @@ typedef struct VabHdr {
     unsigned char  pan;
     unsigned char  attr1;
     unsigned char  attr2;
-    unsigned long  reserved1;
+    u32            reserved1;
 } VabHdr;
 
 typedef struct ProgAtr {
@@ -83,8 +86,8 @@ typedef struct ProgAtr {
     unsigned char mpan;
     char          reserved0;
     short         attr;
-    unsigned long reserved1;
-    unsigned long reserved2;
+    u32           reserved1;
+    u32           reserved2;
 } ProgAtr;
 
 typedef struct VagAtr {
